Add handlers to task_tracker service, add migrations.
